Fix/mapping restrictions (#6)
* Added very simple mapping restrictions.

A source's getter cannot be private and the target needs to be able to see it.
A target's setter cannot be private and the source needs to be able to see it.

Classes marked with the MapperAttribute cannot be an abstract class.

* Changed how it checks if symbols can reach.
It can now check if symbols across assemblies are valid symbols for the purpose they are being checked for.

* Update README.md

* Added restrictions for what the mapper can be created on.
Changed how the generator generates generic methods, according to the new restrictions.