Changed how the attributes works. (#8)
* Changed how the attributes works.
This allows for more flexibility when creating multiple mappers with the same source.
Changed the Standard mapper so it needs to be included with the IncludeAttribute.

Changed so each individual MethodToken contains their own target, rather than a shared target within the ClassToken.

* Begun the process of adding key based attribute functionality to the generator.

* Implemented the start of a key based mapping system.
Begun on the IncludeAttribute.

* Made some minor changes to how the generator gets property values and check for valid attributes.

* Made small changes to some functionality inside the SyntaxProviders and the Generator helpers.

Added the ability to explicitly set the generated name within a IncludeAttribute.

* Working on implementation of reverse mapping

* Implemented MappingStrategy, this could be made better (a lot, but I will look into it when doing a complete refactor of the project again.)

* Changed the applied method in the parameter of the mapper method, to the source of the method.